move extend data
