+++ src/modules/fancyzones/FancyZonesLib/WindowKeyboardSnap.h
@@ -1,9 +1,38 @@
 #pragma once
 
+#include <FancyZonesLib/Zone.h>
+
 class WorkArea;
 
 class WindowKeyboardSnap
 {
+    struct ExtendWindowModeData
+    {
+        HWND window{ nullptr };
+        ZoneIndexSet windowInitialIndexSet{};
+        ZoneIndex windowFinalIndex{ -1 };
+
+        bool IsExtended(HWND wnd) const  
+        {
+            return window == wnd && windowFinalIndex != -1;
+        }
+
+        void Set(HWND w)
+        {
+            window = w;
+            windowFinalIndex = -1;
+            windowInitialIndexSet.clear();
+        }
+
+        void Reset()
+        {
+            window = nullptr;
+            windowFinalIndex = -1;
+            windowInitialIndexSet.clear();
+        }
+    };
+
+
 public:
     WindowKeyboardSnap() = default;
     ~WindowKeyboardSnap() = default;
@@ -14,4 +43,5 @@ private:
     bool SnapHotkeyBasedOnZoneNumber(HWND window, DWORD vkCode, HMONITOR monitor, const std::unordered_map<HMONITOR, std::unique_ptr<WorkArea>>& activeWorkAreas);
     bool SnapHotkeyBasedOnPosition(HWND window, DWORD vkCode, HMONITOR monitor, const std::unordered_map<HMONITOR, std::unique_ptr<WorkArea>>& activeWorkAreas);
     bool ProcessDirectedSnapHotkey(HWND window, DWORD vkCode, bool cycle, WorkArea* const workArea);
+    ExtendWindowModeData m_extendData{}; // Needed for ExtendWindowByDirectionAndPosition
 };
